Remove lidarr type file
