+++ lidarr/system.go
@@ -3,9 +3,39 @@ package lidarr
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"golift.io/starr"
 )
+
+// SystemStatus is the /api/v1/system/status endpoint.
+type SystemStatus struct {
+	Version           string    `json:"version"`
+	BuildTime         time.Time `json:"buildTime"`
+	IsDebug           bool      `json:"isDebug"`
+	IsProduction      bool      `json:"isProduction"`
+	IsAdmin           bool      `json:"isAdmin"`
+	IsUserInteractive bool      `json:"isUserInteractive"`
+	StartupPath       string    `json:"startupPath"`
+	AppData           string    `json:"appData"`
+	OsName            string    `json:"osName"`
+	OsVersion         string    `json:"osVersion"`
+	IsMonoRuntime     bool      `json:"isMonoRuntime"`
+	IsMono            bool      `json:"isMono"`
+	IsLinux           bool      `json:"isLinux"`
+	IsOsx             bool      `json:"isOsx"`
+	IsWindows         bool      `json:"isWindows"`
+	IsDocker          bool      `json:"isDocker"`
+	Mode              string    `json:"mode"`
+	Branch            string    `json:"branch"`
+	Authentication    string    `json:"authentication"`
+	SqliteVersion     string    `json:"sqliteVersion"`
+	MigrationVersion  int64     `json:"migrationVersion"`
+	URLBase           string    `json:"urlBase"`
+	RuntimeVersion    string    `json:"runtimeVersion"`
+	RuntimeName       string    `json:"runtimeName"`
+	StartTime         time.Time `json:"startTime"`
+}
 
 // GetSystemStatus returns system status.
 func (l *Lidarr) GetSystemStatus() (*SystemStatus, error) {
